totalizedFormat in process...
