+++ src/components/Design/DataClient/index.js
@@ -1,0 +1,2 @@
+import { DataClient } from "./DataClient";
+export default DataClient;
