+++ src/components/Design/DataTypeDesign/styles.js
@@ -1,24 +1,9 @@
 import styled from "@emotion/styled";
 import { colors } from "../../../styles/colors";
 
-export const Wrapper = styled.div`
-	display: grid;
-	grid-template-columns: 1fr 1fr;
-	> div{
-		display: flex;
-		flex-direction: column;
-		gap: 6px;
-		width: 90%;
-	}
-	@media(max-width: 1024px){
-		display:flex;
-		flex-direction:column;
-		gap: 0.5rem;
-	}
-`
 
 export const Container = styled.div`
-margin: 10px 80px;
+ margin: 10px 80px;
 
 `
 export const Content = styled.div`
