+++ src/pages/TotalizedFormat/index.js
@@ -1,0 +1,2 @@
+import TotalizedFormat from "./TotalizedFormat";
+export default TotalizedFormat
